+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 - Add config option for setting a maximum number of resident credentials.
+- Ignore public key credential paramters with an unknown type, as required by
+  the Webauthn spec ([#28][])
+
+[#28]: https://github.com/solokeys/fido-authenticator/issues/28
 
 ## [0.1.1] - 2022-08-22
 - Fix bug that treated U2F payloads as APDU over APDU in NFC transport @conorpp